requirements.txt
tensorflow
flask
numpy
pillow
opencv-python
scikit-learn
matplotlib